refactor: delete unneeded files
