+++ packages/runner/pubspec.yaml
@@ -13,11 +13,10 @@ environment:
 dependencies:
   flutter_gen_core: ^4.1.6
   build: '>=2.0.0 <3.0.0'
-  collection: ^1.15.0
+  collection: ^1.16.0
   crypto: ^3.0.2
   glob: ^2.0.2
 
 dev_dependencies:
   flutter_lints: ^2.0.1
   build_test: '>=2.0.0 <3.0.0'
-  flutter_lints: '>=1.0.4 <2.0.0'
